+++ Tests/FluentAssertions.Specs/Collections/CollectionAssertionSpecs.cs
@@ -2455,6 +2455,137 @@ namespace FluentAssertions.Specs
 
         #endregion
 
+        #region Not Contain In Order
+
+        [Fact]
+        public void When_two_collections_contain_the_same_items_but_in_different_order_it_should_not_throw()
+        {
+            // Arrange
+            IEnumerable collection = new[] { 1, 2, 3 };
+
+            // Act / Assert
+            collection.Should().NotContainInOrder(2, 1);
+        }
+
+        [Fact]
+        public void When_a_collection_does_not_contain_an_ordered_item_it_should_not_throw()
+        {
+            // Arrange
+            IEnumerable collection = new[] { 1, 2, 3 };
+
+            // Act / Assert
+            collection.Should().NotContainInOrder(4, 1);
+        }
+
+        [Fact]
+        public void When_a_collection_contains_less_items_it_should_not_throw()
+        {
+            // Arrange
+            IEnumerable collection = new[] { 1, 2 };
+
+            // Act / Assert
+            collection.Should().NotContainInOrder(1, 2, 3);
+        }
+
+        [Fact]
+        public void When_a_collection_does_not_contain_a_range_twice_it_should_not_throw()
+        {
+            // Arrange
+            IEnumerable collection = new[] { 1, 2, 1, 3, 12, 2, 2 };
+
+            // Act / Assert
+            collection.Should().NotContainInOrder(1, 2, 1, 1, 2);
+        }
+
+        [Fact]
+        public void When_asserting_collection_does_not_contain_some_values_in_order_but_collection_is_null_it_should_throw()
+        {
+            // Arrange
+            IEnumerable collection = null;
+
+            // Act
+            Action act = () => collection.Should().NotContainInOrder(4);
+
+            // Assert
+            act.Should().Throw<XunitException>().WithMessage("Cannot verify absence of ordered containment in a <null> collection.");
+        }
+
+        [Fact]
+        public void When_two_collections_contain_the_same_items_in_the_same_order_it_should_throw()
+        {
+            // Arrange
+            IEnumerable collection = new[] { 1, 2, 2, 3 };
+
+            // Act
+            Action act = () => collection.Should().NotContainInOrder(new[] { 1, 2, 3 }, "that's what we expect");
+
+            // Assert
+            act.Should().Throw<XunitException>().WithMessage(
+                "Expected collection {1, 2, 2, 3} to not contain items {1, 2, 3} in order because that's what we expect, " +
+                "but items appeared in order ending at index 3.");
+        }
+
+        [Fact]
+        public void When_collection_contains_contain_the_same_items_in_the_same_order_with_null_value_it_should_throw()
+        {
+            // Arrange
+            IEnumerable collection = new object[] { 1, null, 2, "string" };
+
+            // Act
+            Action act = () => collection.Should().NotContainInOrder(1, null, "string");
+
+            // Assert
+            act.Should().Throw<XunitException>().WithMessage(
+                "Expected collection {1, <null>, 2, \"string\"} to not contain items {1, <null>, \"string\"} in order, " +
+                "but items appeared in order ending at index 3.");
+        }
+
+        [Fact]
+        public void When_the_first_collection_contains_a_duplicate_item_without_affecting_the_order_it_should_throw()
+        {
+            // Arrange
+            IEnumerable collection = new[] { 1, 2, 3, 2 };
+
+            // Act
+            Action act = () => collection.Should().NotContainInOrder(1, 2, 3);
+
+            // Assert
+            act.Should().Throw<XunitException>().WithMessage(
+                "Expected collection {1, 2, 3, 2} to not contain items {1, 2, 3} in order, " +
+                "but items appeared in order ending at index 2.");
+        }
+
+        [Fact]
+        public void When_two_collections_contain_the_same_duplicate_items_in_the_same_order_it_should_throw()
+        {
+            // Arrange
+            IEnumerable collection = new[] { 1, 2, 1, 2, 12, 2, 2 };
+
+            // Act
+            Action act = () => collection.Should().NotContainInOrder(1, 2, 1, 2, 12, 2, 2);
+
+            // Assert
+            act.Should().Throw<XunitException>().WithMessage(
+                "Expected collection {1, 2, 1, 2, 12, 2, 2} to not contain items {1, 2, 1, 2, 12, 2, 2} in order, " +
+                "but items appeared in order ending at index 6.");
+        }
+
+        [Fact]
+        public void When_passing_in_null_while_checking_for_absence_of_ordered_containment_it_should_throw()
+        {
+            // Arrange
+            IEnumerable collection = new[] { 1, 2, 3 };
+
+            // Act
+            Action act = () => collection.Should().NotContainInOrder(null);
+
+            // Assert
+            act.Should().Throw<ArgumentNullException>().WithMessage(
+                "Cannot verify absence of ordered containment against a <null> collection.*");
+        }
+
+        #endregion
+
         #region (Not) be in order
 
         [Fact]